+++ database/migrations/1480234575515_users.js
@@ -1,0 +1,19 @@
+const Schema = use('Schema');
+
+class UsersTableSchema extends Schema {
+
+  up () {
+    this.create('users', (table) => {
+      table.increments();
+      table.string('email', 254).notNullable().unique();
+      table.timestamps();
+    });
+  }
+
+  down () {
+    this.drop('users');
+  }
+
+}
+
+module.exports = UsersTableSchema;